refactor: split SettingsPage.tsx into section components
Extract each self-contained settings section (theme, personalization,
watchlists, AI memory, MCP tokens, data export, daily brief, weekly
recap, privacy, updates, delete account) into its own component under
frontend/src/components/settings/. SettingsPage.tsx is now a thin
layout file that composes them.

Closes #829
